feat: restructure scaffolding
