+++ backend/package.json
@@ -3,8 +3,8 @@
     "version": "1.0.0",
     "main": "index.js",
     "scripts": {
-        "dev": "node index.js --entorno=dev",
-        "migration:run": "sequelize-cli db:migrate"
+        "dev": "node index.js",
+        "migration:run": "sequelize-cli db:migrate --migrations-path=./database/migrations"
     },
     "keywords": [],
     "author": "",
